+++ CHANGELOG.md
@@ -4,12 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Deprecations/Changes: `Alert`: Add `displayIcon` boolean prop, defaults to true. `Alert` now always displays an icon, unless `displayIcon` is defined to `false`.
+- Deprecations/Changes: `Alert`: Replace `iconBadgeBorderColor` prop with `iconHasBorder` boolean prop, defaults to false.
 - Deprecations/Changes: `Alert`: Deprecate the following props:
   - `success`, replaced by `status="success"`,
   - `error`, replaced by `status="danger"`,
   - `warning`, replaced by `status="warning"`
 - Feature: `Alert`: Add `status="info"` default prop.
-- Deprecations/Changes: `Alert`: Remove `iconBadgeBorderColor` prop.
 - Fix: `AlertBox`: Improve styles.
 - Feature: Add the following components:
   - `ChatLayout`